move duration results files to duration/tests/results
