+++ statsmodels/duration/tests/results/phreg_gentests.py
