+++ statsmodels/duration/tests/results/survival.R
