+++ statsmodels/duration/tests/results/survival_r_results.py
